requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
streamlit
pyyaml
plotly
joblib
streamlit